+++ docs/changelog.rst
@@ -23,6 +23,14 @@ In progress (unreleased)
 
 * :gh:issue:`1451` :mod:`mitogen`: Refactor module whitelist & blacklist with
   module overrides and blocks. Improve error messages for denied modules.
+* :gh:issue:`1451` :mod:`mitogen`: Move ``Importer.ALWAYS_BLACKLIST``
+  to :attr:`mitogen.core.ImportPolicy.unsuitables`
+* :gh:issue:`1451` :mod:`mitogen`: Always delegate modules in
+  :attr:`mitogen.core.ImportPolicy.unsuitables` to Python's own importers
+* :gh:issue:`1451` :mod:`mitogen`: Add discovered stdlib module names to
+  :attr:`mitogen.core.ImportPolicy.unsuitables`
+* :gh:issue:`1451` :mod:`mitogen`: Add modules discovered on the controller
+  :attr:`mitogen.core.ImportPolicy.unsuitables`
 
 
 v0.3.41 (2026-02-10)